clean code 1_80
# message auto-generated for no-merge-commit merge:
merge clean_code_1_80 into master

Clean Code 1到80条

Created-by: msprofDeveloper

Match-id-4d257c0977b5267882de3ab3f00d589a8f2f5561

+++ collector/dvvp/msprof/engine/src/aicpu_plugin.cpp
@@ -126,10 +126,10 @@ int AicpuPlugin::ReceiveStreamData(CONST_VOID_PTR data, unsigned int dataLen)
     MSVP_MAKE_SHARED0_RET(fileChunkReq, analysis::dvvp::ProfileFileChunk, PROFILING_FAILED);
     fileChunkReq->fileName = Utils::PackDotInfo(message->filename(), jobCtx.tag);
     fileChunkReq->chunk = std::move(std::string(message->chunk().c_str(), message->chunksizeinbytes()));
-    fileChunkReq->chunkSize = message->chunksizeinbytes();
+    fileChunkReq->chunkSize = static_cast<size_t>(message->chunksizeinbytes());
     fileChunkReq->isLastChunk = message->islastchunk();
     fileChunkReq->chunkModule = message->datamodule();
-    fileChunkReq->offset = message->offset();
+    fileChunkReq->offset = static_cast<size_t>(message->offset());
     fileChunkReq->extraInfo = Utils::PackDotInfo(jobCtx.job_id, logicDevIdStr_);
     fileChunkReq->chunkStartTime = 0U;
     fileChunkReq->chunkEndTime = 0U;
